refactor: move components to folder
