+++ app/profile/_components/index.ts
@@ -1,0 +1,4 @@
+import DeleteUserButton from "./DeleteUserButton";
+import LogoutButton from "./LogoutButton";
+
+export { DeleteUserButton, LogoutButton };
